Analytics Addition
Added ANALYTICS table to capture statistics from the API calls.  Will currently track # of downloads (only if Issue Security is enabled - no way to determine if downloaded directly from Baker without hitting API first) as well as "API Interactions".  This will gather basic statistics on the number of times methods are called on the API for a given Publication.  Data will be eventually presented on a Analytics Dashboard page along with other information.